datepicker: prefer showing as many past dates per month as possible
This small change will really only make a difference when launching
the calendar the first time.  If the user starts manually browsing
to the next/prev months this is kinda moot.  The idea is that it
is more common to choose (recent) past dates over (recent) future
dates.  This isn't true for all apps (booking future reservations)
but is common for most (searching for recent transactions activity,
running last week's report, etc).  Plus on things that involve
choosing future dates, it is more likely you'd have to manually
advance months.

This is all to say that I'm asserting that last few days of the
previous month will more likely chosen over the first few days of
the next month.  The switches the calendar start date to always
prefer showing as many days of the previous month as possible.
On most month, at least one day in the next month will also be
shown but there is a case where the last day of the current month
is a Saturday so no next month days will be shown.

Overall, I think this is a pretty minor change that might have a
slight UX improvement.  Also it bugged me that the previous
implementation would have cases where over a week of next month's
dates were shown.  This seemed silly to me and I always seemed to
look over them bc I would just say "ahh, I'm at the end of the
month - advance to next month".