css: update border radius circle style classes
This updates the circle border radius style classes to use the
`50%` value.  This is closer to the spirit of these style classes.
If your elem is rectangular, the border will be an oval.  If it is
square, the border will be a true circle.

This previous behavior of the circle style classes produced more
of a "pill" shaped border (if rectangular, if square close to a
circle).  To preserve the old behavior and to name it more
accurately, I've added -pill style classes.