indicator text input: update to allow pos on left or right
This updates the indicator text input to allow positioning the
indicator on either the left or the right.  This is prep for
adding a currency text input which will need to place the
currency indicator on the left.

This also cleans up the indicator padding calculations to
detect the indicator elem width so it doesn't have to be manually
set unless you want to use a custom value.  I also made it so you
can customize the padding value to use to account for non-std
borders and such.

I chose to keep the indicator elem positioned on the right by
default but updated the select dropdown and datepicker to manually
set their indicator pos value in case the default changes.