add currency text input component; add RomoCurrency format functions
This adds a new text input component for entering currency values.
This input is an indicator text input that can optionally format
the input on change.  It proxies all indicator text input data
attrs and events as well.

This also adds functions to a new `RomoCurrency` namespace for
formatting and un-formatting currency strings.  These are used
internally by the currency text input but they seemed handy
so I decided to make stand-alone functions out of them.