sortable: add api to manually bind draggable elems
This can be used to re-bind the sortable dragging logic to
draggable elems that have been inline edited or regenerated.
This provides a `doBindDraggableElems` method and a
`sortable:bindDraggableElems` trigger event that takes an elem
collection and re-binds all dragging logic on those elems.

This also reworkes the initial drag binding logic to use the
new method.