(hotfix) select dropdown: force filter text input to size="1"
This is to keep any default input sizing from affecting the width
of the dropdown body/content elems.  By default, most text inputs
are size="20", which means they will be 20 text chars width wide.
If you want a select that is more narrow than 20 text chars, the
input won't allow it and will force the popup width to be wider
than the configured with (which is applied to the dropdown body).

`1` is the smallest width I can set this to (`0` has no effect on
the default width size).  I'm essentially saying the minimum
width a select dropdown can be is the the 1 char width plus
input/popup padding.  I think this will work for almost every
scenario.