break out OptionListDropdown component from SelectDropdown
This is a new component that contains all of the logic related to
a dropdown with a list of selectable options and a filter input
for selecting those options.  It also contains all of the dropdown
handling and navigating prev/next options.

This is prep for adding a new "picker" component.  This new
component will be very similar to the "select" component.  It too
will compose the OptionListDropdown as it will have a list of
selectable options in a dropdown.  The main difference is that
the selectable options won't be sourced from markup but instead
be fetched via RomoAjax as the user types in the filter input.
This component will be designed to work well with large sets of
options that are too cumbersome to render inline as markup.

The OptionListDropdown is designed to allow defining ad-hoc
options via the `doSetListItems` method.  It also triggers a
an event when the filter is typed that can implement custom
behavior.  The SelectDropdown listens for this event and does
its word boundary filter logic.  The picker will listen for this
event and use RomoAjax to fetch matching options and manually
rebuild the option list.