(hotfix) update option list dropdown ex data to include all attrs
In a previous fix commit, I switched up the implementation to
expect all attrs on options.  This updates the ex data comment
to include both `'displayText'` and `'displayHtml'` attrs.
I also updated them to show `'value'` first b/c I think it is
more logical to put this value first so I want the example to
reflect this.