(hotfix) fix a couple of select bugs due to option list dropdown
I didn't test these thoroughly enough.  This moves the optgroup
padding to just the custom dropdown LI and not the native ui.  If
the spacing is on the native UI, the spacing goes around all of
the group's child options and makes it look like some weird dark
border thing.

This also adds setting the preset selected value on selects.  This
was added to the picker but was missed on the select.  This makes
the custom dropdown elem show the preset selected value.