add SelectedOptionsList js component; design picker/select interactions
This is prep for adding multi support for romo pickers/selects.
This component is the common UI handling for the selected options
for each.

This component's api includes methods to set items, clear items,
add item, remove item, and refresh ui.  The design is that you
modify the item set using the api methods and then call refresh
ui to update the markup.

In addtion, each selected item's markup supports an optional
rm icon.  If clicked, the component publishes an rm icon clicked
event that can be bound to for running remove logic.  It is
expected that the rm api method would be called at some point
after the event is published.

Here are some of the other smaller UI features:

* the selected ui is synced on refresh - the items are compared to
  the ui and only the items that need to be added are added and
  the ui items that need removing are removed.  This avoids any
  ui flashes with a clear and rebuild approach.
* the lists height grows as needed up to an option max rows.  If
  no max-rows is configured, the height grows infinitely.
* the component is built with a "focus elem".  This component is
  not designed to be directly focused (it will be paired with a
  select/picker that will be focused).  If the elem is clicked, it
  focuses its focus elem.  If its focus elem is focused/blurred,
  a focus style class is added/removed to the elem to make it
  appear as it is focused/blurred as part of the focus elem being
  focused/blurred.

This is the first pass at this component and its usage. In a
coming, I'll intergrate this with selects/pickers with the multi
prop and polish the UX.

Note: I went ahead and psuedo-code added notes for integrating
with the select/picker.  I wanted to be sure I could use this
as is and integrate it as designed.  These notes will be removed
in a coming effort.