add multi picker support
This updates the picker, similarly to how the select was updated
in PR 126, to support a `multiple` property for picking multiple
values.

As with selects, this integrates the selected options list with
the picker in the 3 main places: setting preset values, setting
newly selected individual items, and removing individual items.

The main difference from the select implementation is getting and
setting the values.  There isn't options markup to updated so
instead it serializes the values with a customizable delim.  To
get the values, split the elem value.  To set them, join the
values and set as the elems value.  Whether setting values in a
multi picker or not, we only ever need to set a single display
text for the dropdown.  If multi, we always set `''`, otherwise
we set the single display text.