(hotfix) update custom event api check for event triggers
IE11 and earlier doesn't fully support the CustomEvent api.  I put
in a polyfill for this but the conditional to detect when the
polyfill wasn't good enough.

This switches to a better condition to check when the polyfill
is needed.

See https://stackoverflow.com/questions/26596123/internet-explorer-9-10-11-event-constructor-doesnt-work
for details and where I got the better conditional from.