(hotfix) fix bug in calculating handling max height in sel opt list
If no max rows param is given, max rows is treated as zero.  This
causes a bad max height value to be calculated.  However, this max
height val should be ignored it no max rows param is given.  The
conditional was checking whether max rows was undefined to know
if it should ignore the max height of not.  However, max rows will
never be undefined - it will either be the given value or zero.
This fixes the conditional.