add base `data` and `setData` methods
This continues the goal of replacing the jQuery api that Romo
uses with its own vanilla js api.  This adds `data` and `setData`
methods to compliment the `attr` and `setAttr` methods.  The
difference is that the data methods auto prepend `data-` to the
given data names and they serialize/deserialize the values the
get/set. The data methods use the attr methods to read/persist
the values from/to the DOM.