romoAjax: fix bug with invokes with data
This adds a second event handler, `onTriggerInvoke` that is now
used exculsively with the 'romoAjax:triggerInvoke' event.  This
allows having separate event processing for triggered invokes
vs auto invokes (from a click for example).  Auto invokes never
expect/take data where trigger invokes do.  This fixes a bug
where you use an event that passes unrelated data as an auto
invoke.  Without this, the unrelated data would be interpreted
as invoke data which is not intended.

This also adds `return false` directives to the picker's romo
ajax event handlers.  Since romo ajax components are prone to
being nested, it is a good practice to stop propagation on any
romo ajax events.  This handles that for pickers.