Update `RomoAjax` component to not use jquery
This updates the `RomoAjax` component to not use jquery. This is
part of the effort to no longer require jquery to use Romo. This
removes all of the jquery usage within the `RomoAjax` component
and also updates the dropdown, modal, picker, and tooltip
components to not use the jquery method for initializing a romo
ajax component.

This removes the romo ajax component unbinding all invoke on
events from its elem when initialized. The `Romo.off` doesn't
support unbinding all events but we also decided that Romo doesn't
need to be this aggressive with its components elems. This allows
users to bind event handlers to the romo ajax's `invokeOn` event
if desired (though it's not recommended).

Note: This didn't update any event names because they are already
prefixed with `romoAjax`.